refactor: extract `store_in_stack` into utils
