+++ src/codegen/operations.rs
@@ -1,15 +1,12 @@
 use melior::{
-    dialect::{
-        arith,
-        llvm::{self, r#type::pointer, LoadStoreOptions},
-    },
+    dialect::arith,
     ir::{Attribute, Block, Location},
     Context as MeliorContext,
 };
 use num_bigint::BigUint;
 
 use super::context::CodegenCtx;
-use crate::{constants::STACK_GLOBAL_VAR, opcodes::Operation, utils::llvm_mlir};
+use crate::{opcodes::Operation, utils::store_in_stack};
 
 pub fn generate_code_for_op(
     context: CodegenCtx,
@@ -30,7 +27,6 @@ fn codegen_push(
 ) -> Result<(), String> {
     let context = &codegen_ctx.mlir_context;
     let location = Location::unknown(context);
-    let ptr_type = pointer(context, 0);
 
     let constant_value = block
         .append_operation(arith::constant(
@@ -42,34 +38,7 @@ fn codegen_push(
         .unwrap()
         .into();
 
-    let stack_baseptr_ptr = block
-        .append_operation(llvm_mlir::addressof(
-            context,
-            STACK_GLOBAL_VAR,
-            ptr_type,
-            location,
-        ))
-        .result(0)
-        .unwrap();
-
-    let stack_baseptr = block
-        .append_operation(llvm::load(
-            context,
-            stack_baseptr_ptr.into(),
-            ptr_type,
-            location,
-            LoadStoreOptions::default(),
-        ))
-        .result(0)
-        .unwrap();
-
-    block.append_operation(llvm::store(
-        context,
-        constant_value,
-        stack_baseptr.into(),
-        location,
-        LoadStoreOptions::default(),
-    ));
+    store_in_stack(context, block, constant_value);
 
     Ok(())
 }
